+++ CHANGELOG.md
@@ -6,6 +6,14 @@ File format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 See cloning and downloading instructions [here](https://www.ivoyager.dev/developers/).
 
+## [v0.0.25] - UNRELEASED
+
+Developed using Godot 4.4.1
+
+### Added
+* IVBodyFinisher class for adding non-procedural nodes from table data (to help declutter IVBody).
+
+
 ## [v0.0.24] - 2025-03-31
 
 Developed using Godot 4.4.
@@ -146,6 +154,7 @@ Requires plugin [ivoyager_table_reader](https://github.com/ivoyager/ivoyager_tab
 ##
 I, Voyager projects v0.0.16 and earlier used a different core submodule [ivoyager](https://github.com/ivoyager/ivoyager) (now depreciated); see previous changelog [here](https://github.com/ivoyager/ivoyager/blob/master/CHANGELOG.md).
 
+[v0.0.25]: https://github.com/ivoyager/ivoyager_core/compare/v0.0.24...HEAD
 [v0.0.24]: https://github.com/ivoyager/ivoyager_core/compare/v0.0.23...v0.0.24
 [v0.0.23]: https://github.com/ivoyager/ivoyager_core/compare/v0.0.22...v0.0.23
 [v0.0.22]: https://github.com/ivoyager/ivoyager_core/compare/v0.0.21...v0.0.22